+++ src/python/rv/chunks/chunk.py
@@ -5,12 +5,14 @@ class Chunk:
 
     chnm = None
     has_chff = False
+    has_chfr = False
 
     def chunks(self):
         yield (b"CHNM", pack("<I", self.chnm))
         yield (b"CHDT", self.chdt())
         if self.has_chff:
             yield (b"CHFF", self.chff())
+        if self.has_chfr:
             yield (b"CHFR", self.chfr())
 
     def chdt(self):