+++ src/python/rv/modules/sampler.py
@@ -3,6 +3,7 @@ from collections import OrderedDict
 from io import BytesIO
 from itertools import chain
 from struct import pack, unpack
+from typing import List, Optional
 
 from logutils import BraceMessage as _F
 from rv.controller import Controller
@@ -11,6 +12,7 @@ from rv.modules import Module
 from rv.modules.base.sampler import BaseSampler
 from rv.note import NOTE
 from rv.readers.reader import read_sunvox_file
+from rv.synth import Synth
 
 log = logging.getLogger(__name__)
 
@@ -31,6 +33,8 @@ class Sampler(BaseSampler, Module):
     options_chnm = 0x0101
 
     behaviors = {B.receives_notes, B.sends_audio}
+
+    effect: Optional[Synth]
 
     class NoteSampleMap(OrderedDict):
         start_note = NOTE.C0
@@ -62,8 +66,8 @@ class Sampler(BaseSampler, Module):
         initial_enable = None
         initial_sustain = None
         initial_loop = None
-        initial_controller_number = 0
-        initial_gain_percentage = 100
+        initial_ctl_index = 0
+        initial_gain_pct = 100
         initial_velocity = 0
         initial_controller = 0
         _legacy_point_bytes = None
@@ -81,8 +85,8 @@ class Sampler(BaseSampler, Module):
             self.enable = self.initial_enable
             self.sustain = self.initial_sustain
             self.loop = self.initial_loop
-            self.controller_number = self.initial_controller_number
-            self.gain_percentage = self.initial_gain_percentage
+            self.ctl_index = self.initial_ctl_index
+            self.gain_pct = self.initial_gain_pct
             self.velocity = self.initial_velocity
             self.loaded = False
 
@@ -121,13 +125,19 @@ class Sampler(BaseSampler, Module):
             data = pack(
                 "<HBBB",
                 self.bitmask,
-                self.controller_number,
-                self.gain_percentage,
+                self.ctl_index,
+                self.gain_pct,
                 self.velocity,
             )
             data += b"\0\0\0"
-            data += pack("<I", len(self.points))
-            data += b"\0\0\0\0\0\0\0\0"
+            data += pack(
+                "<HHHH",
+                len(self.points),
+                self.sustain_point,
+                self.loop_start_point,
+                self.loop_end_point,
+            )
+            data += b"\0\0\0\0"
             for x, y in self.points:
                 y -= self.range[0]
                 data += pack("<HH", x, y)
@@ -136,8 +146,8 @@ class Sampler(BaseSampler, Module):
         def load_chdt(self, chdt):
             (
                 self.bitmask,
-                self.controller_number,
-                self.gain_percentage,
+                self.ctl_index,
+                self.gain_pct,
                 self.velocity,
                 _,
                 _,
@@ -207,7 +217,7 @@ class Sampler(BaseSampler, Module):
         def __init__(self):
             self.data = b""
             self.loop_start = 0
-            self.loop_end = 0
+            self.loop_len = 0
             self.volume = 64
             self.finetune = 100
             self.format = Sampler.Format.float32
@@ -241,6 +251,8 @@ class Sampler(BaseSampler, Module):
     vibrato_rate = Controller((0, 63), 0, attached=False)
     volume_fadeout = Controller((0, 8192), 0, attached=False)
 
+    samples: List[Optional[Sample]]
+
     def __init__(self, **kwargs):
         super(Sampler, self).__init__(**kwargs)
         self.volume_envelope = self.VolumeEnvelope()
@@ -273,10 +285,13 @@ class Sampler(BaseSampler, Module):
             self.effect_control_envelopes[2].chunks(),
             self.effect_control_envelopes[3].chunks(),
         ]
-        if self.effect:
-            iters += self.effect.chunks()
         for iter in iters:
             yield from iter
+        if self.effect:
+            f = BytesIO()
+            self.effect.write_to(f)
+            yield b"CHNM", b"\x0a\1\0\0"
+            yield b"CHDT", f.getvalue()
         yield from super(Sampler, self).specialized_iff_chunks()
         for i, sample in enumerate(self.samples):
             if sample is not None:
@@ -332,7 +347,7 @@ class Sampler(BaseSampler, Module):
         w = f.write
         w(pack("<I", sample.frames))
         w(pack("<I", sample.loop_start))
-        w(pack("<I", sample.loop_end))
+        w(pack("<I", sample.loop_len))
         w(pack("<B", sample.volume))
         w(pack("<b", sample.finetune))
         sustain_flag = 4 if sample.loop_sustain else 0
@@ -417,7 +432,7 @@ class Sampler(BaseSampler, Module):
         sample = self.samples[index] = self.Sample()
         data = chunk.chdt
         (sample.loop_start,) = unpack("<I", data[0x04:0x08])
-        (sample.loop_end,) = unpack("<I", data[0x08:0x0C])
+        (sample.loop_len,) = unpack("<I", data[0x08:0x0C])
         sample.volume = data[0x0C]
         (sample.finetune,) = unpack("<b", data[0x0D:0x0E])
         loop_format_flags = data[0x0E]